add CRTP
