+++ paddle/cinn/ir/group_schedule/search/config_searcher.cc
@@ -17,7 +17,6 @@
 #include "paddle/cinn/ir/group_schedule/search/config_searcher.h"
 #include "paddle/cinn/common/target.h"
 #include "paddle/cinn/ir/group_schedule/config/database.h"
-#include "paddle/cinn/ir/group_schedule/config/filedatabase.h"
 #include "paddle/cinn/utils/string.h"
 
 namespace cinn {
