+++ test/cpp/pir/cinn/file_tile_config_test.cc
@@ -19,7 +19,7 @@
 
 #include "paddle/cinn/hlir/dialect/operator/ir/op_dialect.h"
 #include "paddle/cinn/ir/group_schedule/config/database.h"
-#include "paddle/cinn/ir/group_schedule/config/filedatabase.h"
+#include "paddle/cinn/ir/group_schedule/config/file_database.h"
 #include "paddle/cinn/ir/group_schedule/config/group_tile_config.h"
 #include "paddle/cinn/ir/group_schedule/search/config_searcher.h"
 #include "paddle/cinn/ir/group_schedule/search/measurer.h"
